Filtering complete?
Filter works. Any time a key is lifted in the filter box, the hidden class is activated for every item, and then items matching the filter are unhidden. The filter IS case sensitive right now. Not sure how to fix that.
Filtration matching is done by looking at all spans on the page - Title and Description are the only real elements (other than glyphicons) using span, so it only searches titles and description. If other spans are ever added to part of the idea div, the filtration would be less accurate because it might catch unwanted ideas.