Filter complete
Idea descriptions that are longer than 100 characters are now shortened to 100 characters if there are no spaces or up to 100 characters, stopping at the latest space character that isn't past character 100. The trailing appendix ('...') can be defined in the function as can max length. Added API endpoint, show, which returns a single idea - using this, when you click to edit an existing idea, you can get access to the whole/original description instead of the truncated version that's displayed on the site.
The method for truncating is to count down from character 100 looking for a space. When found, it uses that as the endcount for creating a substring. If the string doesn't have more than 100 characters, it doesn't bother with the work, but it nullifies the trailing characters so they aren't appended on.